AiLearning/pyproject.toml
[tool.poetry]
name = "AiLearning"
version = "0.3.0"
description = "Ejemplos de uso programaticos para aprender AI"
authors = ["Mario Codes <mariocodes1993@gmail.com>"]
readme = "README.md"
packages = [
	{ include = "InsertAndQueryVectorDatabase" },
	{ include = "LangChain" },
	{ include = "Rag" },
	{ include = "Testing" },
]

[tool.poetry.dependencies]
python = "^3.14"
python-dotenv = "^1.2.1"
dotenv = "^0.9.9"
pandas = "^2.3.3"
openai = "^2.11.0"
langchain-openai = "^1.1.7"
langchain = "^1.2.3"
azure-cosmos = "^4.15.0"
pypdf = "^6.7.5"
langchain-community = "^0.4.1"
langchain-text-splitters = "^1.1.1"

[tool.poetry.scripts]
call-to-openai = "Testing.CallToOpenAI:main"
manually-get-embeddings = "Testing.ManualCallToGetEmbeddingsFromText:main"

insert-manual-embeddings = "InsertAndQueryVectorDatabase.InsertEmbeddings:main"
query-manual-embeddings = "InsertAndQueryVectorDatabase.QueryVectorialDB:main"

insert-rag-files = "Rag.ChunkFilesIntoCosmos:main"
ask-rag-question = "Rag.AskWithRagContext:main"

langchain = "LangChain.LangChain:main"
langchain-lcel = "LangChain.LangChainExpressionLanguage:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
